listo

+++ client_crud.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, jsonify, request
-
+from validar_key import Validar_Token
 from conexion import create_db_connection
 
 clients_bp = Blueprint('clients', __name__)
@@ -7,6 +7,10 @@ clients_bp = Blueprint('clients', __name__)
 # Clientes
 @clients_bp.route('/clientes', methods=['GET'])
 def get_clients():
+        # Se valida la API Key enviada en la cabecera del request
+    api_key = request.headers.get('autorizacion')
+    if not Validar_Token(api_key):
+        return jsonify({'mensaje': 'API Key inválida'}), 401
     connection = create_db_connection()
     cursor = connection.cursor(dictionary=True)
     query = "SELECT * FROM clientes"
@@ -18,6 +22,10 @@ def get_clients():
 
 @clients_bp.route('/clients/<int:client_id>', methods=['GET'])
 def get_client(client_id):
+            # Se valida la API Key enviada en la cabecera del request
+    api_key = request.headers.get('autorizacion')
+    if not Validar_Token(api_key):
+        return jsonify({'mensaje': 'API Key inválida'}), 401
     connection = create_db_connection()
     cursor = connection.cursor(dictionary=True)
     query = "SELECT * FROM clients WHERE id=%s"
@@ -29,6 +37,10 @@ def get_client(client_id):
 
 @clients_bp.route('/clients', methods=['POST'])
 def create_client():
+            # Se valida la API Key enviada en la cabecera del request
+    api_key = request.headers.get('autorizacion')
+    if not Validar_Token(api_key):
+        return jsonify({'mensaje': 'API Key inválida'}), 401
     client = request.json
     connection = create_db_connection()
     cursor = connection.cursor()
@@ -41,6 +53,10 @@ def create_client():
 
 @clients_bp.route('/clients/<int:client_id>', methods=['PUT'])
 def update_client(client_id):
+            # Se valida la API Key enviada en la cabecera del request
+    api_key = request.headers.get('autorizacion')
+    if not Validar_Token(api_key):
+        return jsonify({'mensaje': 'API Key inválida'}), 401
     client = request.json
     connection = create_db_connection()
     cursor = connection.cursor()
@@ -53,6 +69,10 @@ def update_client(client_id):
 
 @clients_bp.route('/clients/<int:client_id>', methods=['DELETE'])
 def delete_client(client_id):
+            # Se valida la API Key enviada en la cabecera del request
+    api_key = request.headers.get('autorizacion')
+    if not Validar_Token(api_key):
+        return jsonify({'mensaje': 'API Key inválida'}), 401
     connection = create_db_connection()
     cursor = connection.cursor()
     query = "DELETE FROM clients WHERE id=%s"
